feat: automatic down-leveling of declarations files (#11)
In order to ensure compatibility of output `.d.ts` files with previous
compiler release lines, we automatically emit down-leveled `.d.ts` files
and insert `typesVersions` settings in `package.json` appropriately.

This also adds a new phase to the integration test that validates the
`1.x` release line of the `jsii` compiler can compile projects that
depend on packages built by the current one, and which make use of
TypeScript-3.9-unsupported features.

Closes aws/jsii#3501

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0