chore: add test for auto-accessors
Auto accessor support was introduced in TypeScript 4.9 and allows
introducing accessor-implemented properties (backed by an implicit
`#private` property) using the `accessor` keyword. This is not
compatible with the use of `readonly` (such properties are always
read-write), and is useful to provide implementation of accessor-based
abstract properties, or to override accessor-based parent members.