chore: add test for auto-accessors (#22)
Auto accessor support was introduced in TypeScript 4.9 and allows
introducing accessor-implemented properties (backed by an implicit
`#private` property) using the `accessor` keyword. This is not
compatible with the use of `readonly` (such properties are always
read-write), and is useful to provide implementation of accessor-based
abstract properties, or to override accessor-based parent members.



---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0