feat: prohibit use of separate write type on properties (#18)
Separate write types are allowed on property setters since TypeScript
4.3, where a setter can accept more types than the getter's return type,
granted the getter's return type is assignable to the setter's argument.
This is however not a feature that can be modeled in the jsii type
system, or that can be accurately represented in several target
languages (e.g: C#).



---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0