fix: #private members not ignored (#19)
ECMAScript `#private` identifiers are "truly" private (in contrast with
how the TypeScript `private` keyword just hides declarations away from a
type's declarations).

TypeScript does not support introducing these with the `private` keyword
as it would be aking to a stutter. However, the `Assembler` needs to
correctly detect private identifiers & treat them as private.

Also modified the `jsii-calc` suite to include several such members as a
test that these are properly ignored.



---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0