fix: correctly reject index signatures (#20)
Index signatures cannot be represented in the jsii type model and should
be rejected. TypeScript 4.3 also introduces support for static index
signatures, which should be rejected, too.

Starting with TypeScript 4.4, Symbols can be used as keys on index
signatures. Neither symbols nor index signatures are supported, but only
the later should be reported (to avoid double-reporting issues).

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>