fix: do not pass semver objects to downlevel-dts
If `downlevel-dts` has a different `semver` dependency instance than the
jsii compiler, an `instanceof` check internal to `semver` fails and
produces a `TypeError` exception. We can easily avoid this by always
providing the target version as a `string`, which leads to
`downlevel-dts` ultimately using its own `semver` instance to turn it
into the `semver.SemVer` object it needs.