feat: allow marking indexers @internal (#38)
There may be internal use-cases for declaring indexer properties, and
those can be erased using the `@internal` JSDoc tag. This allows using
such declarations while explicitly hiding them from the library's API.

Additionally improved error messaging to remove related information when
it points to types in the TypeScript standard library, as those are
often fairly complex and their declaration is not expected to be shown
to users.



---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0