feat: add "@jsii ignore" directive (#42)
Users can annotate declarations with `@jsii ignore` to explicitly opt
them out of cross-language support. This still allows the declaration to
be visible to Javascript & TypeScript consumers of the library, but it
is not included in the jsii assembly, and is hence not available for use
in other languages.

Example usage:
```ts
export interface StructWithExtraProperties {
  readonly propertyA: string;
  readonly propertyB: number;

  /**
   * The index signature is not supported by jsii across languages. This
   * allows it to be used in TypeScript and Javascript, but the library
   * author should have an alternate (perhaps less convenient or less
   * TypeScript-idiomatic) way to achieve the same intent for other
   * languages.
   *
   * @jsii ignore
   */
  readonly [key: string]: string | number;
}
```

This is semantically similar to `@internal` except it does not hide the
declaration from Javascript consumers.



---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0