fix: deepo-import of index files breaks ts3.9 compilers
The `typesVersions` directive applies recursively when the TypeScript
compiler searches for the appropriate object to load... When attempting
to resolve a directory, it will first rewrite that directory according
to the typesVersions directive, and then it will rewrite the `index`
sub-path within there again, resulting in two distinct rewrites
happening within the same resolution. This appears to be a TypeScript
bug as reported in microsoft/TypeScript#43133.

The work-around is to include a second rewrite candidate that explicitly
targets the `*/index.d.ts` sub-path, which makes removes the need for a
rewrite when searching down the `index` route.

Causes projen/projen#2570