fix: stop down-leveling files where only formatting has changed
Use a TypeScript-aware normalization on the declarations files to
neutralize any formatting discrepancies that may exist between the
output of `downlevel-dts` and the output of the raw TypeScript compiler.

Also corrects the location of output files in case where `tsc.outDir`
is configured (previously a spurious interim path was added).