fix: stop down-leveling files where only formatting has changed (#102)
Use a TypeScript-aware normalization on the declarations files to
neutralize any formatting discrepancies that may exist between the
output of `downlevel-dts` and the output of the raw TypeScript compiler.

Also corrects the location of output files in case where `tsc.outDir` is
configured (previously a spurious interim path was added).

Finally, changes how the down-leveled declarations are created when
necessary to avoid having to retain all their text in memory, as this
can result in a very large amount of data being retained in working set.
(Fixes #40)

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0