feat: allow specifying a deprecation message regardless of stability (#146)
Previously, `jsii` required the package's stability to be `deprecated`
if there cas a `deprecated` message configured. This is however
problematic when a `stable` package reaches End-of-Support and the
maintainer wants to signal this in npmjs.com.

Instead of failing, only log a warning to possibly raise awareness on an
unintended configuration.

Foward-ports: aws/jsii#4145



---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0