feat: export common functions as a submodule (#188)
Backport of #187 

The `symbolIdentifier` function is a base jsii feature that other tools
can depend on without having to depend on the rest of jsii. Currently a
dependant package has to import all of the jsii code just to get
`symbolIdentifier`. This can have performance implications, as well as
producing larger bundles.

This change exports `symbolIdentifier` as part of a new `jsii/common`
submodule to allow a direct import.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0