chore: fix upgrade-dependencies task (#247)
We do not use the `exclude` keyword and likely will not for the
foreseeable future. Just dropping it for now. We will have to
re-implement that if we need it but I don't see why we would -- this
decision was also made in projen:
https://github.com/projen/projen/pull/2845/files. This continues work
started because projen had breaking changes.


---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0

---------

Signed-off-by: github-actions <github-actions@github.com>
Signed-off-by: Kaizen Conroy <36202692+kaizencc@users.noreply.github.com>
Co-authored-by: github-actions <github-actions@github.com>