feat: user-provided TypeScript config (#931)
This PR introduces support for user-provided TypeScript configs through
new options to the CLI, the programmatic API and the `jsii` config
section in `package.json` respectively:

| CLI | package.json | Programmatic API | Description |
|-----|--------------|-------------------|-------------|
| `-c`, `--tsconfig` | `jsii.tsconfig` | `typeScriptConfig` | Use this
typescript configuration file to compile the jsii project. |
| `--validate-tsconfig` | `jsii.validateTsConfig` |
`validateTypeScriptConfig` | Validate the provided typescript
configuration file against a set of rules.|

The following validation rule sets are available:

| Rule set | Description |
|---------|-------------|
| **strict** ⭐ | Validates the provided config against a strict rule set
designed for maximum backwards-compatibility. |
| **generated** | Enforces a config as created by `--generate-tsconfig`.
Use this to stay compatible with the generated config, but have full
ownership over the file. |
| **minimal** | Only enforce options that are known to be incompatible
with jsii. This rule set is likely to be incomplete and new rules will
be added without notice as incompatibilities emerge. |
| **off** 🧪 | Disables all config validation, including options that are
known to be incompatible with jsii. Intended for experimentation only.
Use at your own risk. |

Resolves https://github.com/aws/jsii/issues/2071

---

## Refactors

Mostly changed the compiler by moving some code around or into new
locations to be reused. Specifically the generated config was created in
multiple places without re-use. With this PR, loading or building the
tsconfig now happens once in the constructor and this config is used
through out.


## Testing strategy

Compiler changes are covered by existing test cases.
Added property-based tests for the new config validator.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0