chore: enable backports of PRs to previous version lines (#976)
Because this project is using a custom release branches component, we
have to instantiate the `PullRequestBackport` component manually (as
opposed to project options).

Includes a couple of refactors to make working with `SupportPolicy`
easier, these are copied from jsii-rosetta.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0