chore: fix failing tsconfig validator prop test (#983)
There was a proptest that failed in our test pipeline:

```
FAIL test/tsconfig/validator.test.ts
  ● Built-in matchers › Match.arrEq › pass

    Property failed after 63 tests
    { seed: -1990728707, path: "62:2:64", endOnFailure: true }
    Counterexample: [[[Number.NaN],[Number.NaN]]]
```
To fix this, I refactored the `looseEqual` helper. This also made some
of the matcher code simpler.
Also added a few missing tests for some matchers.

In practice this is probably not very relevant, so I made this a chore.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0