chore: only publish jsii metrics on main branch (#1057)
This PR adds a conditional to only publish the JSII Benchmark metrics to
CloudWatch when on the main branch. We do not want to publish metrics on
every PR build, as these metrics are used to trigger an alarm if the
JSII slowdown passes the threshold.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0

---------

Signed-off-by: Sumu <sumughan@amazon.com>