chore: fix ci PRs might get merged with failing tests (#1210)
We had a PR https://github.com/aws/jsii-compiler/pull/1191 that got
merged despite a failure in Node22 tests.

This happened because on a failure of one of the matrix builds, the job
that is unifying them is skipped. This is documented behavior:

![image](https://github.com/user-attachments/assets/c4eb7ae6-1b37-4f5e-91c4-65a5e5661514)

The fix is to make sure the combination step always runs and manually
check the status of the matrix build.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0