fix: typesVersions not respecting outDir from tsconfig file (#1238)
Fixes #1234 

Also fixes an issue related the `.type-compat` output path that was
discovered when adding tests. When getting `outdir` from the program,
this might be either a relative or absolute path. Previously the code
assumed it was always relative which is not correct. This led to
absolute paths being used for the `.type-compat` dir and in the post
processing; ultimately leading to a compilation failure. The fix is to
ensure `outdir` & `.type-compat` are normalized to a relative path.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0