feat(tsconfig): mark verbatimModuleSyntax, isolatedModules & isolatedDeclarations options as configurable (#1262)
The `isolatedModules` & `isolatedDeclarations` options increase the
strictness of the TypeScript compiler.

The `verbatimModuleSyntax` option is a modern replacement for a set of
older options that influence if unused imports a
persevered in the JS output. As such it is irrelevant to the jsii type
system and at the discretion of the user.

No testing required, as these are covered by the existing proptest.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0

(cherry picked from commit c36b4ee42579a513aac768c27c19c0d6a8fbf02e)