fix: omitting optional property fails C# compilation
The following is a minimal reproducing example:

```ts
export interface ISomeInterface {
  readonly xyz?: string;
}

export abstract class SomeClass implements ISomeInterface {
}
```

This breaks code generation for C#, as the `SomeClass._Proxy` class
we generate will try to generate an `override` for the `Xyz` property,
which is missing from the `SomeClass` class itself.

This issue is exhibited by a difference in logic between how we iterate
the members of a class when we generate the class itself vs when we
generate its proxy.

However, after looking at the code I'm not convinced it's correct either
way (nor for Java), so to be safe we picked the solution where we force
the user to declare the property on the class.

The correct declaration would be:

```ts
export abstract class SomeClass implements ISomeInterface {
  public abstract readonly xyz?: string;
}
```