chore: update integ test tarball to be compatible with ts5.6 (#1469)
#1468 was merged despite failing benchmark tests. 

This was due to a new TS compiler error on the old code. I've fix this
issue by manually updating the offending code in the `aws-cdk-lib`
tarball. Done this so the benchmark will stay comparable for the v5.6 to
changes.


---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0