fix: `lib` setting from custom config is ignored
When a user provides a custom `tsconfig.json` file, they have the ability
to override the `lib` setting, which provides a number of libraries
that should be loaded at compile time.

This config setting is not passed directly into the TypeScript compiler.
Instead, it is resolved and translated into a list of `.d.ts` file that
are passed as `rootNames` to the compiler, along with the actual source
files.

In that resolution, we failed to look at the `lib` setting the user
provided, and instead always reverted to the default config.

Fix by looking at the user-provided config, falling back to the default
config only if the user config doesn't have a `lib` setting (mirroring
the behavior of tsc).

Fixes #4706.