fix: `lib` setting from custom config is ignored (#1576)
When a user provides a custom `tsconfig.json` file, they have the
ability to override the `lib` setting, which provides a number of
libraries that should be loaded at compile time.

This config setting is not passed directly into the TypeScript compiler.
Instead, it is resolved and translated into a list of `.d.ts` file that
are passed as `rootNames` to the compiler, along with the actual source
files.

In that resolution, we failed to look at the `lib` setting the user
provided, and instead always used the default config.

Fix by looking at the user-provided config, falling back to the default
config only if the user config doesn't have a `lib` setting (mirroring
the behavior of tsc).

Fixes aws/jsii#4706.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0