feat: intersection types (#2325)
This adds support for intersection types to the jsii compiler.

There are a number of restrictions to the use of intersection types:

- They can only appear in argument position, or in structs that are
exclusively used as the input to functions.
    - Not as return values of functions
    - Not in structs that are returned by functions
    - Not as arguments to constructors
    - Not as property types
- They must intersect only (behavioral) interface types.
    - Not structs
    - Not primitives
    - Not class types
- All shared members of interfaces that are intersected must have the
same types and arguments as each other.



---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>