fix: invalid jsii target parameters are not validated (#2398)
Arguments to the jsii `targets` array that did not make sense in the
target language were not being validated, leading to confusing
compile-time errors.

Add validation for a bunch of these.


---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0