fix: submodule config for files is erroneously loaded from directory
When a submodule is exported from a file in a directory, we look for
`.jsiirc.json` in the same directory.

That was only intended for exporting a `directory/index.ts` as
a submodule, and is wrong when submodules are exported from files
in the same directory. Consider:

```
- root.ts                 export * from './submodule';
- submodule
    +-- index.ts          export * from './subsubmodule';
    +-- subsubmodule.ts
    +-- .jsiirc.json     // <-- gets used for both submodules!
```

Instead, only use the directory-level `.jsiirc.json` if the exported
file is `index.ts`. Otherwise, for submodules read the jsii config from
`subsubmodule.jsiirc.json`.